add: node title
